backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Document Processing
docling==1.16.2
PyPDF2==3.0.1
pdf2image==1.17.0
pypdf==4.0.1

# OCR Engines
pytesseract==0.3.10
easyocr==1.7.1
paddleocr==2.7.3
rapidocr-onnxruntime==1.3.11

# AI/LLM
google-generativeai==0.3.2

# Image Processing
Pillow==10.2.0
opencv-python==4.9.0.80

# Database & Storage
firebase-admin==6.4.0
supabase==2.3.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
aiofiles==23.2.1
celery==5.3.6
redis==5.0.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0
